Moved files into the challenge package
