--- src/main/java/com/cegeka/academy/service/ChallengeService.java
@@ -1,5 +1,0 @@
-package com.cegeka.academy.service;
-
-public interface ChallengeService {
-
-}
