+++ src/main/java/com/cegeka/academy/service/challenge/ChallengeService.java
@@ -1,0 +1,5 @@
+package com.cegeka.academy.service.challenge;
+
+public interface ChallengeService {
+
+}
